Move the SanitizeFragment tests to FragmentTests.cs
They stay members of HtmlSanitizerTests, declared partial across the two
files, rather than moving to a class of their own. ThreadTest reflects over
typeof(HtmlSanitizerTests) to decide what to run concurrently, so tests in
a separate class would quietly drop out of that sweep - which is precisely
what backs the thread safety SanitizeFragment is documented as having.

ThreadTestDiscoversTestsAndTheoryRowsTest now asserts one of the moved
tests is discovered, so undoing the arrangement fails rather than silently
shrinking the sweep. The nameof makes that a compile error.
